+++ modules/service/src/main/scala/ch/epfl/bluebrain/nexus/kg/service/io/BaseEncoder.scala
@@ -1,21 +1,11 @@
 package ch.epfl.bluebrain.nexus.kg.service.io
 
 import akka.http.scaladsl.model.Uri
-import ch.epfl.bluebrain.nexus.kg.service.hateoas.Links
 import ch.epfl.bluebrain.nexus.kg.service.config.Settings.PrefixUris
-import ch.epfl.bluebrain.nexus.kg.service.io.BaseEncoder.JsonLDKeys._
+import ch.epfl.bluebrain.nexus.kg.service.io.RoutesEncoder.JsonLDKeys._
 import io.circe._
-import io.circe.syntax._
 
 class BaseEncoder(prefixes: PrefixUris) {
-
-  implicit val linksEncoder: Encoder[Links] =
-    Encoder.encodeJson.contramap { links =>
-      links.values.mapValues {
-        case href :: Nil => Json.fromString(s"$href")
-        case hrefs       => Json.arr(hrefs.map(href => Json.fromString(s"$href")): _*)
-      }.asJson
-    }
 
   /**
     * Syntax to extend JSON objects in response body with the various contexts used internally.
@@ -24,14 +14,23 @@ class BaseEncoder(prefixes: PrefixUris) {
     */
   implicit class JsonOps(json: Json) {
 
+    /**
+      * Adds or merges the core context URI to an existing JSON object.
+      */
     def addCoreContext: Json = addContext(prefixes.CoreContext)
 
+    /**
+      * Adds or merges the links context URI to an existing JSON object.
+      */
     def addLinksContext: Json = addContext(prefixes.LinksContext)
 
+    /**
+      * Adds or merges the search context URI to an existing JSON object.
+      */
     def addSearchContext: Json = addContext(prefixes.SearchContext)
 
     /**
-      * Adds or merges the standard context URI to an existing JSON object.
+      * Adds or merges a context URI to an existing JSON object.
       *
       * @param context the standard context URI
       * @return a new JSON object
@@ -59,22 +58,4 @@ class BaseEncoder(prefixes: PrefixUris) {
       }
     }
   }
-}
-
-object BaseEncoder {
-
-  object JsonLDKeys {
-    val `@context`     = "@context"
-    val `@id`          = "@id"
-    val links          = "links"
-    val resultId       = "resultId"
-    val source         = "source"
-    val score          = "score"
-    val nxvNs          = "nxv"
-    val nxvRev         = "nxv:rev"
-    val nxvDeprecated  = "nxv:deprecated"
-    val nxvDescription = "nxv:description"
-    val nxvPublished   = "nxv:published"
-  }
-
 }
